refactor(PeerRoom): move to lib
