+++ src/components/Room/usePeerVerification.ts
@@ -2,7 +2,7 @@ import { useCallback, useContext, useEffect, useState } from 'react'
 import { ShellContext } from 'contexts/ShellContext'
 import { Peer, PeerVerificationState } from 'models/chat'
 import { encryption } from 'services/Encryption'
-import { PeerRoom } from 'services/PeerRoom'
+import { PeerRoom } from 'lib/PeerRoom'
 import { PeerActions } from 'models/network'
 import { verificationTimeout } from 'config/messaging'
 import { usePeerNameDisplay } from 'components/PeerNameDisplay'
